# lib/api-zod/src/generated/types/associateProgress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PhaseProgress } from './phaseProgress';

export interface AssociateProgress {
  associateId: number;
  totalCheckins?: number;
  completedCheckins?: number;
  overallPercent?: number;
  phases: PhaseProgress[];
}