# lib/api-zod/src/generated/types/associateUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AssociateUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  cohortId?: number;
  /** @nullable */
  directorId?: number | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  teamLeadId?: number | null;
  /** @nullable */
  seniorMentorId?: number | null;
  position?: string;
  department?: string;
}