# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Associate } from './associate';
import type { CheckinPhase } from './checkinPhase';
import type { CheckinStatus } from './checkinStatus';
import type { CheckinSupervisorRole } from './checkinSupervisorRole';
import type { Supervisor } from './supervisor';

export interface Checkin {
  id: number;
  associateId: number;
  supervisorId: number;
  supervisorRole: CheckinSupervisorRole;
  phase: CheckinPhase;
  status: CheckinStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: Date | null;
  createdAt: Date;
  supervisor?: Supervisor;
  associate?: Associate;
}