# lib/api-zod/src/generated/types/checkinCompletion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinCompletion {
  notes: string;
}