# lib/api-zod/src/generated/types/cohortInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CohortInput {
  /** @minLength 1 */
  name: string;
  /** Calendar date YYYY-MM-DD */
  startDate: string;
  description?: string;
}