# lib/api-zod/src/generated/types/phaseProgress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Checkin } from './checkin';
import type { PhaseProgressPhase } from './phaseProgressPhase';

export interface PhaseProgress {
  phase: PhaseProgressPhase;
  completedCount?: number;
  totalCount?: number;
  checkins: Checkin[];
}