# lib/api-zod/src/generated/types/supervisor.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Associate Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SupervisorRole } from './supervisorRole';

export interface Supervisor {
  id: number;
  name: string;
  email: string;
  role: SupervisorRole;
  /** @nullable */
  title?: string | null;
  createdAt: Date;
}